requirements.txt
amqp==2.1.4
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
boto==2.43.0
certifi==2017.4.17
configparser==3.5.0
contextlib2==0.5.4
coverage==4.3.1
decorator==4.0.10
diff-match-patch==20121119
dj-database-url==0.4.1
Django==1.9.8
django-admin-rangefilter==0.3.0
django-admin-sortable2==0.6.18
django-allauth==0.28.0
django-anymail==0.6.1
django-appconf==1.0.2
django-autofixture==0.12.1
django-ckeditor==5.4.0
django-colorfield==0.1.10
django-compat==1.0.13
django-crispy-forms==1.7.0
django-debug-toolbar==1.6
django-easy-select2==1.3.3
django-filer==1.2.5
django-flat-responsive==1.0.0
django-flat-theme==1.1.4
django-hijack==2.1.2
django-hijack-admin==2.1.2
django-import-export==0.5.1
django-js-asset==0.1.1
django-materializecss-form==1.0.1
django-mptt==0.8.6
django-pkgconf==0.2.0
django-polymorphic==1.0.2
django-redis-cache==1.7.1
django-robots==2.0
Django-Select2==5.8.10
django-storages-redux==1.3.2
django-summernote==0.8.7
django-versatileimagefield==1.6
django-widget-tweaks==1.4.1
djangocms-admin-style==1.2.6.2
easy-thumbnails==2.3
enum34==1.1.6
Faker==0.7.3
flake8==3.2.1
gunicorn==19.6.0
holidays==0.8.1
httplib2==0.9.2
ipaddress==1.0.17
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
kombu==4.0.2
mccabe==0.5.3
mixer==5.6.6
oauthlib==2.0.0
opbeat==3.5.2
packaging==16.8
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
psycopg2==2.7.1
ptyprocess==0.5.1
py==1.4.32
pycodestyle==2.2.0
pyflakes==1.3.0
Pygments==2.1.3
pyowm==2.6.0
pyparsing==2.2.0
pytest==3.0.5
pytest-cov==2.4.0
pytest-django==3.1.2
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2016.10
raven==5.32.0
redis==2.10.5
requests==2.12.1
requests-oauthlib==0.7.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
tablib==0.11.4
traitlets==4.3.1
twilio==5.6.0
Unidecode==0.4.19
urllib3==1.21.1
vine==1.1.3
wcwidth==0.1.7
whitenoise==3.2
